商品削除機能修正1

+++ app/controllers/items_controller.rb
@@ -35,9 +35,13 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if item = Item.find(params[:id])
+      item.user_id == current_user.id
+      item.destroy
+      redirect_to root_path
+    else
+      redirect_to root_path
+    end
   end
 
   private
